Extract grid generation, add note to problems.py, simplify k-means++.
The problems.py note remarks that using mixed integer programming with
HCKM (with SCIP as the solver) leads to a slowly converging bound even
though it finds the best solution pretty quickly.

And I found out through experimentation that the best way to give
the MIP solvers candidate points by k-means++ seeding is to just
generate as many points as there are candidate districts, directly.